+++ examples/express-ts/paths/_index.ts
@@ -4,6 +4,7 @@ import { fetchApi } from './fetch-api'
 import { guard } from './guard'
 import { hello } from './hello'
 import { helloName } from './hello/name'
+import { paginationTest } from './pagination-test'
 import { requestValidation } from './request-validation'
 import { validation } from './validation'
 
@@ -12,6 +13,7 @@ export const paths = [
 	customResolver,
 	fetchApi,
 	guard,
+	paginationTest,
 	hello,
 	helloName,
 	requestValidation,